requirements.txt
torch
torchvision
# pip install torch==1.13.1+cu117 torchvision==0.14.1+cu117 torchaudio==0.13.1 --extra-index-url https://download.pytorch.org/whl/cu117
tensorboard     # torch<1.3 requires additional package 'future'
Pillow==9.5.0   # 1865: Pillow==6.2.2
pyyaml
termcolor
prettytable
matplotlib
tqdm
wandb
optuna
# chardet

einops    # if you tend to use the ViT model in template project, please install this package
